[HOTSPOT-346] refactor: prepare, commit step 디렉토리 구조 리팩토링
